requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.1
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
cytoolz==0.12.1
distlib==0.3.2
Django==4.1.7
django-crispy-forms==2.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.0.12
fonttools==4.38.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-resources==5.12.0
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
kiwisolver==1.4.4
lru-dict==1.1.8
matplotlib==3.7.0
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
numpy==1.21.4
packaging==23.0
pandas==1.3.4
parsimonious==0.8.1
Pillow==9.4.0
platformdirs==2.2.0
protobuf==3.19.5
pycryptodome==3.17
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2021.3
requests==2.28.2
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
typing_extensions==4.5.0
uniswap-python==0.6.2
urllib3==1.26.14
varint==1.0.2
virtualenv==20.7.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
zipp==3.14.0
